Update README.md
This repository contains a lot of useful content that may be nice for new members to come across. Calling it "archived" risks throwing away years of good work that went into making a training program that (although is not necessary), would be good for members looking to learn more to peruse.